layout 리팩토링 높이 조절
