+++ src/app/(main)/page.tsx
@@ -37,6 +37,17 @@ export default function Home (): React.ReactElement {
       <p>메인 페이지 내용입니다</p>
       <p>메인 페이지 내용입니다</p>
       <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
+      <p>메인 페이지 내용입니다</p>
     </main>
   )
 }
