+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import '../styles/globals.css'
 import React from 'react'
 import blogMetaData from '@/metadata/blogMetaData'
 import UiContextProvider from '@/components/ui-layout/UiContextProvider'
-import { DynamicLayout } from '@/components/ui-layout/main/DynamicLayout'
+import { TapRouteMain } from '@/components/ui-layout/tap-system/TapRouteMain'
 import { AuthSessionProvider } from '@/components/auth/AuthSessionProvider'
 import SideBarProvider from '@/components/ui-layout/sidebar/SiderBarProvider'
 
@@ -58,10 +58,10 @@ export default function RootLayout ({ children }: { children: React.ReactNode })
       <AuthSessionProvider>
         <SideBarProvider>
 
-          <DynamicLayout
+          <TapRouteMain
             page={children}
           >
-          </DynamicLayout>
+          </TapRouteMain>
         </SideBarProvider>
       </AuthSessionProvider>
     </UiContextProvider>
