+++ src/app/memo/[id]/page.tsx
@@ -1,10 +1,13 @@
 import MemoEditor from '@/components/memo/memo_editor/MemoEditor'
 import React from 'react'
+import MemoFolderContainer from '@/components/memo/MemoFolderContainer'
 
 export default async function Page ({ params }: { params: { id: string } }): Promise<React.ReactElement> {
   return (
-    <main className={'flex-grow'}>
-      <MemoEditor pageMemoNumber={params.id}/>
-    </main>
+    <MemoFolderContainer>
+      <main className={'flex-grow'}>
+        <MemoEditor pageMemoNumber={params.id}/>
+      </main>
+    </MemoFolderContainer>
   )
 }
