+++ src/components/ui-layout/tap-system/AppPageRenderer.tsx
@@ -1,4 +1,3 @@
-import MemoFolderContainer from '@/components/memo/MemoFolderContainer'
 import React from 'react'
 import { AuthSessionContext } from '@/components/auth/AuthSessionProvider'
 import SignInAndOutContainer from '@/components/auth/SignInAndOutContainer'
@@ -12,43 +11,34 @@ export function RenderPage ({ tabs, path, page }: {
 }): React.ReactElement | null {
   const { session }: { session: Session } = React.useContext(AuthSessionContext)
 
-  console.log('렌더링 횟수 테스트 중입니다.')
-  console.log(tabs, path, page)
-
   function renderMemoContainer (): React.ReactElement {
     if (session == null) {
       return (
-        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto">
+        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto min-h-full">
           <SignInAndOutContainer/>
         </div>
       )
     } else if (session?.roles.values().next().value !== 'ADMIN') {
       return (
-        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto">
+        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto min-h-full">
           <AdminAccessDenied/>
         </div>
       )
     } else {
       return (
-        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto">
-          <MemoFolderContainer>
-            {page}
-          </MemoFolderContainer>
+        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto min-h-full">
+          {page}
         </div>
       )
     }
   }
 
   function renderOthers (): React.ReactElement | null {
-    if (tabs.length > 0 && (path !== '/empty')) {
-      return (
-        <div className="bg-gray-700 p-4 rounded-b-lg overflow-auto">
-          {page}
-        </div>
-      )
-    } else {
-      return null
-    }
+    return (
+      <div className="bg-gray-700 p-4 rounded-b-lg">
+        {page}
+      </div>
+    )
   }
 
   if (tabs.length > 0 && (path !== '/empty') && path.startsWith('/memo')) {
