+++ src/components/ui-layout/tap-system/hook/useTabs.ts
@@ -1,7 +1,25 @@
 import type React from 'react'
 import { useCallback, useEffect, useState } from 'react'
-import { restoreTabsFromLocalStorage } from '@/components/ui-layout/main/utils/restoreTabs'
-import { type Tab } from '@/components/ui-layout/tapbar/TabItem'
+import { type Tab } from '@/components/ui-layout/tap-system/TabItem'
+
+const restoreTabsFromLocalStorage = (path: string): any => {
+  const savedTabs = localStorage.getItem('tabs')
+  const tabsList = (savedTabs != null) ? JSON.parse(savedTabs) : null
+  if (path.startsWith('/login/oauth2')) {
+    if (tabsList == null || tabsList.length === 0) {
+      return [{ name: '/', context: '/' }]
+    } else {
+      return tabsList
+    }
+  }
+  if (tabsList == null || (tabsList.length === 0 && path !== '/empty')) {
+    return path === '/empty' ? [] : [{ name: path, context: path }]
+  }
+  if ((tabsList.length === 0 && path === '/empty') || (tabsList.length !== 0 && path === '/empty')) {
+    return []
+  }
+  return tabsList
+}
 
 export function useTabs (initialPath: string): {
   tabs: Tab[]
